add log

+++ sdk/appcenter-crashes/src/main/java/com/microsoft/appcenter/crashes/utils/ErrorLogHelper.java
@@ -575,9 +575,10 @@ public class ErrorLogHelper {
      */
     public static void cleanPendingMinidumpDirectory() {
         File[] files = getPendingMinidumpDirectory().listFiles();
-        if (files != null) {
+        if (files != null && files.length > 0) {
             for (File file : files) {
-                file.delete();
+                boolean isDelete = file.delete();
+                AppCenterLog.debug(Crashes.LOG_TAG, String.format("Minidump file '%s' has delete status: '$s'", file.getPath(), isDelete));
             }
         }
     }
